# flask/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "API to store safe multisig transactions"
  version: "v1"
  title: "Gnosis Safe Transaction Service API"
  contact:
    email: "safe@gnosis.io"
  license:
    name: "MIT License"
host: "safe-transaction.mainnet.gnosis.io"
basePath: "/api/v1"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- Basic: []
paths:
...
  /safes/{address}/:
    get:
      tags:
      - "safes"
      description: "Get status of the safe"
      operationId: "safes_read"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/SafeInfoResponse"
        "404":
          description: "Safe not found"
        "422":
          description: "code = 1: Checksum address validation failed\ncode = 50: Cannot\
            \ get Safe info"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/all-transactions/:
    get:
      tags:
      - "safes"
      description: "Returns a paginated list of transactions for a Safe. The list\
        \ has different structures depending on the\ntransaction type:\n- Multisig\
        \ Transactions for a Safe. `tx_type=MULTISIG_TRANSACTION`. If the query parameter\
        \ `queued=False` is\nset only the transactions with `safe nonce < current\
        \ Safe nonce` will be displayed. By default, only the\n`trusted` transactions\
        \ will be displayed (transactions indexed, with at least one confirmation\
        \ or proposed\nby a delegate). If you need that behaviour to be disabled set\
        \ the query parameter `trusted=False`\n- Module Transactions for a Safe. `tx_type=MODULE_TRANSACTION`\n\
        - Incoming Transfers of Ether/ERC20 Tokens/ERC721 Tokens. `tx_type=ETHEREUM_TRANSACTION`"
      operationId: "safes_all_transactions_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "ordering"
        in: "query"
        description: "Which field to use when ordering the results."
        required: false
        type: "string"
      - name: "limit"
        in: "query"
        description: "Number of results to return per page."
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "The initial index from which to return the results."
        required: false
        type: "integer"
      - name: "queued"
        in: "query"
        description: "If `True` transactions with `nonce >= Safe current nonce` are\
          \ also returned"
        required: false
        type: "boolean"
        default: true
      - name: "trusted"
        in: "query"
        description: "If `True` just trusted transactions are shown (indexed, added\
          \ by a delegate or with at least one confirmation)"
        required: false
        type: "boolean"
        default: true
      responses:
        "200":
          description: "A list with every element with the structure of one of these\
            \ transactiontypes"
          schema:
            $ref: "#/definitions/_AllTransactionsSchema"
        "422":
          description: "code = 1: Checksum address validation failed"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/balances/:
    get:
      tags:
      - "safes"
      description: "Get balance for Ether and ERC20 tokens"
      operationId: "safes_balances_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "trusted"
        in: "query"
        description: "If `True` just trusted tokens will be returned"
        required: false
        type: "boolean"
        default: false
      - name: "exclude_spam"
        in: "query"
        description: "If `True` spam tokens will not be returned"
        required: false
        type: "boolean"
        default: false
      responses:
        "200":
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SafeBalanceResponse"
        "404":
          description: "Safe not found"
        "422":
          description: "Safe address checksum not valid"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/balances/usd/:
    get:
      tags:
      - "safes"
      description: "Get balance for Ether and ERC20 tokens with USD fiat conversion"
      operationId: "safes_balances_usd_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "trusted"
        in: "query"
        description: "If `True` just trusted tokens will be returned"
        required: false
        type: "boolean"
        default: false
      - name: "exclude_spam"
        in: "query"
        description: "If `True` spam tokens will not be returned"
        required: false
        type: "boolean"
        default: false
      responses:
        "200":
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SafeBalanceUsdResponse"
        "404":
          description: "Safe not found"
        "422":
          description: "Safe address checksum not valid"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/collectibles/:
    get:
      tags:
      - "safes"
      description: "Get collectibles (ERC721 tokens) and information about them"
      operationId: "safes_collectibles_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "trusted"
        in: "query"
        description: "If `True` just trusted tokens will be returned"
        required: false
        type: "boolean"
        default: false
      - name: "exclude_spam"
        in: "query"
        description: "If `True` spam tokens will not be returned"
        required: false
        type: "boolean"
        default: false
      responses:
        "200":
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SafeCollectibleResponse"
        "404":
          description: "Safe not found"
        "422":
          description: "Safe address checksum not valid"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/creation/:
    get:
      tags:
      - "safes"
      description: "Get status of the safe"
      operationId: "safes_creation_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/SafeCreationInfoResponse"
        "404":
          description: "Safe creation not found"
        "422":
          description: "Owner address checksum not valid"
        "503":
          description: "Problem connecting to Ethereum network"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/delegates/:
    get:
      tags:
      - "safes"
      description: "Get the list of delegates for a Safe address"
      operationId: "safes_delegates_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "Number of results to return per page."
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "The initial index from which to return the results."
        required: false
        type: "integer"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/inline_response_200_3"
        "400":
          description: "Invalid data"
        "422":
          description: "Invalid Ethereum address"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
    post:
      tags:
      - "safes"
      description: "Create a delegate for a Safe address with a custom label. Calls\
        \ with same delegate but different label or\nsigner will update the label\
        \ or delegator if different.\nFor the signature we are using TOTP with `T0=0`\
        \ and `Tx=3600`. TOTP is calculated by taking the\nUnix UTC epoch time (no\
        \ milliseconds) and dividing by 3600 (natural division, no decimals)\nFor\
        \ signature this hash need to be signed: keccak(address + str(int(current_epoch\
        \ // 3600)))\nFor example:\n     - we want to add the owner `0x132512f995866CcE1b0092384A6118EDaF4508Ff`\
        \ and `epoch=1586779140`.\n     - TOTP = epoch // 3600 = 1586779140 // 3600\
        \ = 440771\n     - The hash to sign by a Safe owner would be `keccak(\"0x132512f995866CcE1b0092384A6118EDaF4508Ff440771\"\
        )`"
      operationId: "safes_delegates_create"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "data"
        required: true
        schema:
          $ref: "#/definitions/SafeDelegate"
      responses:
        "202":
          description: "Accepted"
        "400":
          description: "Malformed data"
        "422":
          description: "Invalid Ethereum address/Error processing data"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/delegates/{delegate_address}/:
    delete:
      tags:
      - "safes"
      description: "Delete a delegate for a Safe. Signature is built the same way\
        \ that for adding a delegate.\nCheck `POST /delegates/`"
      operationId: "safes_delegates_delete"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "delegate_address"
        in: "path"
        required: true
        type: "string"
      responses:
        "204":
          description: "Deleted"
        "400":
          description: "Malformed data"
        "404":
          description: "Delegate not found"
        "422":
          description: "Invalid Ethereum address/Error processing data"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/incoming-transactions/:
    get:
      tags:
      - "safes"
      description: "Returns the history of a multisig tx (safe)"
      operationId: "safes_incoming_transactions_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "_from"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "block_number"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "block_number__gt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "block_number__lt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "execution_date__gte"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "execution_date__lte"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "execution_date__gt"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "execution_date__lt"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "to"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "token_address"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "transaction_hash"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "value"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "value__gt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "value__lt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "limit"
        in: "query"
        description: "Number of results to return per page."
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "The initial index from which to return the results."
        required: false
        type: "integer"
      responses:
        "200":
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TransferResponse"
        "422":
          description: "Safe address checksum not valid"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
  /safes/{address}/incoming-transfers/:
    get:
      tags:
      - "safes"
      description: "Returns the history of a multisig tx (safe)"
      operationId: "safes_incoming_transfers_list"
      parameters:
      - name: "address"
        in: "path"
        required: true
        type: "string"
      - name: "_from"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "block_number"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "block_number__gt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "block_number__lt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "execution_date__gte"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "execution_date__lte"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "execution_date__gt"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "execution_date__lt"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "to"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "token_address"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "transaction_hash"
        in: "query"
        description: ""
        required: false
        type: "string"
      - name: "value"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "value__gt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "value__lt"
        in: "query"
        description: ""
        required: false
        type: "number"
      - name: "limit"
        in: "query"
        description: "Number of results to return per page."
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "The initial index from which to return the results."
        required: false
        type: "integer"
      responses:
        "200":
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TransferResponse"
        "422":
          description: "Safe address checksum not valid"
      x-swagger-router-controller: "swagger_server.controllers.safes_controller"
...
securityDefinitions:
  Basic:
    type: "basic"
definitions:
...
  SafeInfoResponse:
    type: "object"
    required:
    - "address"
    - "fallbackHandler"
    - "masterCopy"
    - "modules"
    - "nonce"
    - "owners"
    - "threshold"
    - "version"
    properties:
      address:
        type: "string"
        title: "Address"
      nonce:
        type: "integer"
        title: "Nonce"
      threshold:
        type: "integer"
        title: "Threshold"
      owners:
        type: "array"
        items:
          type: "string"
      masterCopy:
        type: "string"
        title: "Master copy"
      modules:
        type: "array"
        items:
          type: "string"
      fallbackHandler:
        type: "string"
        title: "Fallback handler"
      version:
        type: "string"
        title: "Version"
        minLength: 1
    example:
      address: "address"
      threshold: 6
      owners:
      - "owners"
      - "owners"
      masterCopy: "masterCopy"
      nonce: 0
      version: "version"
      modules:
      - "modules"
      - "modules"
      fallbackHandler: "fallbackHandler"
...
  SafeBalanceResponse:
    type: "object"
    required:
    - "balance"
    - "token"
    - "tokenAddress"
    properties:
      tokenAddress:
        type: "string"
        title: "Token address"
        minLength: 1
      token:
        $ref: "#/definitions/Erc20Info"
      balance:
        type: "string"
        title: "Balance"
        minLength: 1
    example:
      tokenAddress: "tokenAddress"
      balance: "balance"
      token:
        symbol: "symbol"
        decimals: 0
        name: "name"
        logoUri: "logoUri"
...
  inline_response_200_3:
    type: "object"
    required:
    - "count"
    - "results"
    properties:
      count:
        type: "integer"
      next:
        type: "string"
        format: "uri"
        x-nullable: true
      previous:
        type: "string"
        format: "uri"
        x-nullable: true
      results:
        type: "array"
        items:
          $ref: "#/definitions/SafeDelegateResponse"
    example:
      next: "http://example.com/aeiou"
      previous: "http://example.com/aeiou"
      count: 0
      results:
      - delegate: "delegate"
        delegator: "delegator"
        label: "label"
      - delegate: "delegate"
        delegator: "delegator"
        label: "label"
...